+++ packages/riverpod/lib/src/core/element.dart
@@ -31,7 +31,7 @@ base mixin _ProviderRefreshable<OutT, InT> implements Refreshable<OutT> {
 void Function()? debugCanModifyProviders;
 
 /// A Future-like that support synchronous completion.
-@internal
+@publicInMisc
 @publicInCodegen
 typedef WhenComplete = void Function(void Function() cb)?;
 
@@ -615,12 +615,11 @@ depending on itself.
   /// to dependencies that are no-longer used.
   void _performRebuild() {
     runOnDispose();
-    final ref =
-        this.ref = $Ref(
-          this,
-          isFirstBuild: false,
-          isReload: _didChangeDependency,
-        );
+    final ref = this.ref = $Ref(
+      this,
+      isFirstBuild: false,
+      isReload: _didChangeDependency,
+    );
     final previousValue = value;
 
     if (kDebugMode) _debugDidSetState = false;
@@ -684,25 +683,34 @@ depending on itself.
     $Ref<StateT, ValueT> ref,
   );
 
-  /// A utility for re-initializing a provider when needed.
+  var _isFlushing = false;
+
+  /// Recomputes the state of the provider if necessary.
   ///
-  /// Calling [flush] will only re-initialize the provider if it needs to rerun.
-  /// This can involve:
+  /// This is called when:
+  /// - a provider is read
+  /// - a provider is listened
   /// - a previous call to [Ref.invalidateSelf]
   /// - a dependency of the provider has changed (such as when using [Ref.watch]).
   ///
   /// This is not meant for public consumption. Public API should hide
   /// [flush] from users, such that they don't need to care about invoking this function.
   void flush() {
-    if (!_didMount) {
-      _didMount = true;
-      mount();
-    }
+    final wasFlushing = _isFlushing;
+    _isFlushing = true;
+    try {
+      if (!_didMount) {
+        _didMount = true;
+        mount();
+      }
 
-    _maybeRebuildDependencies();
-    if (_mustRecomputeState) {
-      _mustRecomputeState = false;
-      _performRebuild();
+      _maybeRebuildDependencies();
+      if (_mustRecomputeState) {
+        _mustRecomputeState = false;
+        _performRebuild();
+      }
+    } finally {
+      _isFlushing = wasFlushing;
     }
   }
 
@@ -818,7 +826,16 @@ The provider ${_debugCurrentlyBuildingElement!.origin} modified $origin while bu
 
     runOnDispose();
     mayNeedDispose();
-    container.scheduler.scheduleProviderRefresh(this);
+
+    if (
+    // It is redundant to request for a refresh during flushing as the rebuild
+    // will happen immediately
+    !_isFlushing
+        // Don't schedule refresh for paused providers. Those are paused afterall!
+        &&
+        isActive) {
+      container.scheduler.scheduleProviderRefresh(this);
+    }
 
     // We don't call this._markDependencyMayHaveChanged here because we voluntarily
     // do not want to set the _dependencyMayHaveChanged flag to true.
@@ -1001,17 +1018,16 @@ The provider ${_debugCurrentlyBuildingElement!.origin} modified $origin while bu
   }
 
   void addDependentSubscription(ProviderSubscriptionImpl<Object?> sub) {
-    assert(
-      !sub.isPaused && sub.impl.active,
-      'Expected subscription to be active and not paused',
-    );
-
     _onChangeSubscription(sub, () {
+      sub._attachedToElement = true;
+
       if (sub.weak) {
         weakDependents.add(sub);
       } else {
         final dependents = this.dependents ??= [];
         dependents.add(sub);
+
+        if (sub.isPaused || !sub.impl.active) pausedActiveSubscriptionCount++;
       }
 
       if (sub.source case ProviderNode(:final element)) {
@@ -1019,6 +1035,17 @@ The provider ${_debugCurrentlyBuildingElement!.origin} modified $origin while bu
         subs.add(sub);
       }
     });
+
+    // The provider was initialized with a paused subscription. As such, we
+    // need to immediately call onCancel to respect life-cycles.
+    if (!sub.weak &&
+        !_didCancelOnce &&
+        !isActive &&
+        (sub.isPaused || !sub.impl.active)) {
+      _didCancelOnce = true;
+      _runCallbacks(container, ref?._onCancelListeners);
+      onCancel();
+    }
   }
 
   void removeDependentSubscription(