+++ packages/riverpod/lib/src/core/provider/notifier_provider.dart
@@ -36,13 +36,14 @@ typedef RunNotifierBuild<NotifierT, CreatedT> = CreatedT Function(Ref ref, Notif
 /// ```dart
 /// mixin MyMixin<T> extends AnyNotifier<T, FutureOr<T>> {
 ///   @override
-///   void runBuild() {
+///   WhenComplete runBuild() {
 ///     // It is safe to use "ref" here.
 ///     listenSelf((prev, next) => print("New state $next"));
 ///
 ///     print('Before build');
-///     super.runBuild();
+///     final result = super.runBuild();
 ///     print('After build');
+///     return result;
 ///   }
 /// }
 /// ```
@@ -145,10 +146,11 @@ abstract class AnyNotifier<StateT, ValueT> {
   /// ```dart
   /// mixin LoggingMixin<T> on AnyNotifier<T> {
   ///   @override
-  ///   void runBuild() {
+  ///   WhenComplete runBuild() {
   ///     print('Will build $this');
-  ///     super.runBuild();
+  ///     final result = super.runBuild();
   ///     print('Did build $this');
+  ///     return result;
   ///   }
   /// }
   /// ```
@@ -516,16 +518,15 @@ abstract class $ClassProviderElement<
     // ignore: not public
     $Ref<StateT, ValueT> ref,
   ) {
-    final result =
-        classListenable.result ??= $Result.guard(() {
-          final notifier = provider.create();
-          if (notifier._element != null) {
-            throw StateError(alreadyInitializedError);
-          }
+    final result = classListenable.result ??= $Result.guard(() {
+      final notifier = provider.create();
+      if (notifier._element != null) {
+        throw StateError(alreadyInitializedError);
+      }
 
-          notifier._element = this;
-          return notifier;
-        });
+      notifier._element = this;
+      return notifier;
+    });
 
     WhenComplete whenComplete;
     switch (result) {