+++ packages/riverpod_analyzer_utils/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.0-dev.11 - 2026-07-27
+
+- Upgraded `analyzer` to `<15.0.0`
+
 ## 1.0.0-dev.10 - 2026-04-23
 
 Support analyzer 12